requirements.txt
google-generativeai>=0.8.0
elevenlabs>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.20.0
beautifulsoup4>=4.12.0
requests>=2.31.0
feedparser>=6.0.10
PyPDF2>=3.0.0
